backend/scheduled-reminders/requirements.txt
psycopg2-binary==2.9.9
pywebpush==2.0.0
cryptography==41.0.7
